requirements.txt
Appium_Python_Client==2.4.0
pytest==7.1.2
selenium==4.2.0
